fix(missions): 로딩 시 box UI 가 깨지는 문제 수정
