+++ src/components/Home/MissionCardWithSuspense.tsx
@@ -1,14 +1,10 @@
 import { Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
-import { MissionCard } from './MissionCard';
 import { LoadingSpinner, LoadingSpinnerWrapper } from '@/components/common/LoadingSpinner';
 import ErrorFallback from '@/components/common/ErrorFallback';
+import MissionCardList from './MissionCardList';
 
-interface MissionCardWithSuspenseProps {
-  onClick: () => void;
-}
-
-export function MissionCardWithSuspense({ onClick }: MissionCardWithSuspenseProps) {
+export function MissionCardWithSuspense() {
   return (
     <ErrorBoundary
       FallbackComponent={() => <ErrorFallback message="미션을 불러오는데 실패했습니다." />}
@@ -16,11 +12,11 @@ export function MissionCardWithSuspense({ onClick }: MissionCardWithSuspenseProp
       <Suspense
         fallback={
           <LoadingSpinnerWrapper>
-            <LoadingSpinner />
+            <LoadingSpinner size={30} />
           </LoadingSpinnerWrapper>
         }
       >
-        <MissionCard onClick={onClick} />
+        <MissionCardList />
       </Suspense>
     </ErrorBoundary>
   );
