+++ src/pages/Home/Home.tsx
@@ -2,7 +2,7 @@ import { CharacterSection } from '@/components/Home/CharacterSection';
 import { DiaryCard } from '@/components/Home/DiaryCard';
 import { GreetingSection } from '@/components/Home/GreetingSection';
 import { HomeContainer, ActionsSection } from '@/components/Home/Home.styles';
-import { MissionCardWithSuspense } from '@/components/Home/MissionCardWithSuspense';
+import MissionCard from '@/components/Home/MissionCard';
 import { ROUTES } from '@/constants/routes';
 import { useNavigate } from 'react-router-dom';
 
@@ -15,7 +15,7 @@ function Home() {
       <CharacterSection onClick={() => router(ROUTES.CHARACTER)} />
       <ActionsSection>
         <div>
-          <MissionCardWithSuspense onClick={() => router(ROUTES.MISSIONS)} />
+          <MissionCard />
         </div>
         <DiaryCard onClick={() => router(ROUTES.DIARIES)} />
       </ActionsSection>
